changes from pages style route handlers to App router route handlers & added new helper functions to support the modularity to the codebase.
