--- src/app/api/login.ts
@@ -1,35 +1,0 @@
-import bcrypt from "bcryptjs";
-import jwt from "jsonwebtoken";
-import type { NextApiRequest, NextApiResponse } from "next";
-import { PrismaClient } from "../../generated/prisma";
-
-const prisma = new PrismaClient();
-const JWT_SECRET = process.env.JWT_SECRET || "devsecret";
-
-export default async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse,
-) {
-  if (req.method === "POST") {
-    const { email, password } = req.body;
-    if (!email || !password) {
-      return res.status(400).json({ error: "Email and password required" });
-    }
-    const user = await prisma.user.findUnique({ where: { email } });
-    if (!user) {
-      return res.status(401).json({ error: "Invalid credentials" });
-    }
-    const valid = await bcrypt.compare(password, user.password);
-    if (!valid) {
-      return res.status(401).json({ error: "Invalid credentials" });
-    }
-    // Return JWT token for authenticated requests
-    const token = jwt.sign({ id: user.id, email: user.email }, JWT_SECRET, {
-      expiresIn: "2h",
-    });
-    return res.status(200).json({ token, id: user.id, email: user.email });
-  } else {
-    res.setHeader("Allow", ["POST"]);
-    res.status(405).end(`Method ${req.method} Not Allowed`);
-  }
-}
