--- src/app/api/register.ts
@@ -1,35 +1,0 @@
-import bcrypt from "bcryptjs";
-import type { NextApiRequest, NextApiResponse } from "next";
-import { PrismaClient } from "../../generated/prisma";
-
-const prisma = new PrismaClient();
-
-export default async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse,
-) {
-  if (req.method === "POST") {
-    const { email, password } = req.body;
-    if (!email || !password) {
-      return res.status(400).json({ error: "Email and password required" });
-    }
-    // Only allow emails from @oriental.ac.in
-    if (!/^\w+@oriental\.ac\.in$/i.test(email)) {
-      return res
-        .status(400)
-        .json({ error: "Only college emails (@oriental.ac.in) are allowed." });
-    }
-    const existing = await prisma.user.findUnique({ where: { email } });
-    if (existing) {
-      return res.status(409).json({ error: "User already exists" });
-    }
-    const hashed = await bcrypt.hash(password, 10);
-    const user = await prisma.user.create({
-      data: { email, password: hashed },
-    });
-    return res.status(201).json({ id: user.id, email: user.email });
-  } else {
-    res.setHeader("Allow", ["POST"]);
-    res.status(405).end(`Method ${req.method} Not Allowed`);
-  }
-}
